requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
dotenv==0.9.9
fastapi==0.115.0
google-auth==2.47.0
google-genai==1.60.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.38.6
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
